// multisto/src/Schema.ts
export interface Currency {
  ID: number;
  Country?: string | null;
  Currency?: string | null;
  Abbreviation?: string | null;
  Symbol?: string | null;
  /**  Defaults to: 0. */
  isBlockchainBased?: number | null;
  /**  Defaults to: 1. */
  blockchainID: number;
  /**  Defaults to: 0. */
  isNative: number;
  /**  Defaults to: 0. */
  Address: string;
  /**  Defaults to: 0. */
  cryptoReceivingAddress: string;
}
// ...
export interface DividendInvestorPayouts {
  id: number;
  investorId: number;
  payoutId: number;
  amount: number;
  investorShares: number;
  lastUpdatedAt?: Date | null;
  status:
    | "future"
    | "pending"
    | "completed"
    | "rejected"
    | "reverting"
    | "reverted"
    | "exception";
}
// ...
export interface DividendPayouts {
  id: number;
  templateId: number;
  dateTimeFrom: Date;
  dateTimeDue: Date;
  lastUpdatedAt?: Date | null;
  status:
    | "future"
    | "pending"
    | "completed"
    | "rejected"
    | "reverting"
    | "reverted"
    | "exception";
  totalAmount: number;
  companyShares: number;
  totalInvestorShares: number;
  investors?: number | null;
  awardValue?: number | null;
  /**  Defaults to: 0.0000000000000000. */
  currentlyPaidAmount: number;
  /**  Defaults to: 0. */
  currentlyPaidInvestors: number;
  adminId?: number | null;
}
// ...
export interface DividendPayoutsView {
  /**  Defaults to: 0. */
  id: number;
  templateId: number;
  dateTimeFrom: Date;
  dateTimeDue: Date;
  lastUpdatedAt?: Date | null;
  status:
    | "future"
    | "pending"
    | "completed"
    | "rejected"
    | "reverting"
    | "reverted"
    | "exception";
  totalAmount: number;
  companyShares: number;
  totalInvestorShares: number;
  investors?: number | null;
  stoId: number;
  shareTypeId?: number | null;
  currencyId?: number | null;
  channelId?: number | null;
  isActive: number;
  periodUnit: "days" | "months";
  period: number;
  awardValue?: number | null;
  title?: string | null;
  awardStrategy:
    | "formula"
    | "feePerShare"
    | "percentPremiumValuePerShare"
    | "dividendAmountDistributed";
  /**  Defaults to: unused. */
  templateStatus: "unused" | "used" | "historical";
  adminId?: number | null;
}
// ...
export interface DividendTemplates {
  id: number;
  historyId?: number | null;
  stoId: number;
  shareTypeId?: number | null;
  currencyId?: number | null;
  channelId?: number | null;
  isActive: number;
  periodUnit: "days" | "months";
  period: number;
  awardValue: number;
  title?: string | null;
  awardStrategy:
    | "formula"
    | "feePerShare"
    | "percentPremiumValuePerShare"
    | "dividendAmountDistributed";
  awardFormula?: string | null;
  /**  Defaults to: unused. */
  status: "unused" | "used" | "historical";
}
// ...
export interface Paymentchannels {
  ID: number;
  stoid?: number | null;
  paymentType?: number | null;
  title?: string | null;
  paymentDetails?: string | null;
  currencyID?: number | null;
  /**  Defaults to: 1. */
  isActive?: number | null;
  /**  Defaults to: 0. */
  conversionEnabled?: number | null;
  /**  Defaults to: 0. */
  currencyToConvert?: number | null;
  /**  Defaults to: 0.0000000000000000. */
  conversionRate?: number | null;
  /**  Defaults to: 0. */
  canWithdrawFunds?: number | null;
  /**  Defaults to: 0. */
  sendInstructionalDepositEmail?: number | null;
  depositInstructionText?: string | null;
  depositInstructionEmailHeader?: string | null;
  adminEmailHeader?: string | null;
  adminEmailBody?: string | null;
  /**  Defaults to: 1. */
  sendAdminEmail: number;
}
// ...
export interface Sharetypes {
  ID: number;
  title: string;
  stoid: number;
  /**  Defaults to: 0.000. */
  totalShares?: number | null;
  /**  Defaults to: 0.000. */
  companyShares?: number | null;
  /**  Defaults to: 0.0000000000000000. */
  nominalValue?: number | null;
  /**  Defaults to: 0. */
  isNominalValueApplicable: number;
  /**  Defaults to: 0. */
  isVotingRightsApplicable: number;
  /**  Defaults to: 0. */
  isDividendRightsApplicable: number;
  /**  Defaults to: 0. */
  isblockchain: number;
  ethereumContractAddress: string;
  ethereumWhitelistAddress: string;
  /**  Defaults to: 0.0000000000000000. */
  premimum?: number | null;
  currencyid: number;
  /**  Defaults to: 1. */
  needauthorization: number;
  token_abi?: string | null;
  whitelist_abi?: string | null;
  ethereumBlockchainPublicAddress?: string | null;
  /**  Defaults to: default. */
  subscriptionform?: string | null;
  /**  Defaults to: 0.0000. */
  minimumSharesToBuyByInvestor?: number | null;
  /**  Defaults to: 0. */
  blockchainProtocol?: number | null;
  /**  Defaults to: 0. */
  blockchainBuyOrdersAllowed?: number | null;
  /**  Defaults to: 0.000. */
  reduceSharesForPurchase?: number | null;
  /**  Defaults to: 1. */
  isEnabled?: number | null;
  /**  Defaults to: 0. */
  walletCustodayType?: number | null;
  tanganyWalletID?: string | null;
  /**  Defaults to: 0. */
  investorCanPurchaseDirectly?: number | null;
  AssetName?: string | null;
  AssetTag?: string | null;
  /**  Defaults to: 1.00. */
  votingPower: number;
  /**  Defaults to: 1. */
  isMeetingRightsApplicable?: number | null;
  /**  Defaults to: 1. */
  isInvestorTradable?: number | null;
  /**  Defaults to: 18. */
  blockchainDecimals?: number | null;
  ipfsDocumentHash?: string | null;
  /**  Defaults to: 0.000. */
  custodianShares: number;
  /**  Defaults to: 1. */
  sellToCompany: number;
  /**  Defaults to: 0.0000000000000000. */
  sellValue: number;
  /**  Defaults to: 0. */
  isShareNosApplicable: number;
  /**  Defaults to: 0. */
  isCertificateNosApplicable: number;
  /**  Defaults to: -1. */
  channelIDForAutoPayments?: number | null;
  keyStoreFileAutoPayments?: string | null;
  keyStoreFileAutoPaymentsPassword?: string | null;
}